feat: Added main logic & timers
-Added Atmega32 datasheet .pdf
-Setup Timer Driver with .h & .c including APIs and implementation
-Added Interrupt registers in registers.h
-Configured interrupts with sei() , and setup macros for INT0
APP_start() :
-Constructed main logic for application
APP_init() :
-Added needed initialization for interrupts and timers

application.c :
-Added ISR for __vector_1

+++ On-demand Traffic light control/On-demand Traffic light control/ECUAL/Timer Driver/timer.h
@@ -13,7 +13,7 @@
 
 
 void TIMER_init();
-void TIMER_delay(uint32_t millisec);
+void TIMER_delay(uint16_t millisec);
 
 
 #endif /* TIMER_H_ */
